Extract fade animation to @hedvig-ui
